requirements.txt
streamlit==1.31.0
mysql-connector-python>=8.0
python-dotenv>=1.0
pandas==2.1.4
